bluetooth: audio: Add API to get MICP service ATT handles
This is needed for upper tester.
